feat(x/gamm): split pool-related business logic * add single-token deposit methods * TODO: exit 안만듬
